fix bug

+++ base/src/world/chunk.rs
@@ -30,7 +30,7 @@ impl Chunk {
 
     /// Creates a chunk from a `Vec<HexPillar>`
     pub fn from_pillars(pillars: Vec<HexPillar>) -> Self {
-        assert_eq!(pillars.len(), CHUNK_SIZE.pow(2));
+        assert_eq!(pillars.len() as usize, CHUNK_SIZE.pow(2) as usize);
 
         Chunk { pillars: pillars }
     }
